Move StructuralKey to the local execution module
