+++ runners/direct-java/src/main/java/org/apache/beam/runners/direct/DirectExecutionContext.java
@@ -23,6 +23,7 @@ import org.apache.beam.runners.core.StepContext;
 import org.apache.beam.runners.core.TimerInternals;
 import org.apache.beam.runners.direct.WatermarkManager.TimerUpdate;
 import org.apache.beam.runners.direct.WatermarkManager.TransformWatermarks;
+import org.apache.beam.runners.local.StructuralKey;
 
 /**
  * Execution Context for the {@link DirectRunner}.
