+++ runners/local-java/src/test/java/org/apache/beam/runners/local/StructuralKeyTest.java
@@ -16,19 +16,15 @@
  * limitations under the License.
  */
 
-package org.apache.beam.runners.direct;
+package org.apache.beam.runners.local;
 
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.not;
-import static org.hamcrest.Matchers.theInstance;
 import static org.junit.Assert.assertArrayEquals;
-import static org.junit.Assert.assertThat;
 
-import org.apache.beam.runners.local.StructuralKey;
 import org.apache.beam.sdk.coders.ByteArrayCoder;
 import org.apache.beam.sdk.coders.StringUtf8Coder;
 import org.apache.beam.sdk.coders.VarIntCoder;
 import org.hamcrest.Matchers;
+import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
@@ -40,8 +36,9 @@ import org.junit.runners.JUnit4;
 public class StructuralKeyTest {
   @Test
   public void getKeyEqualToOldKey() {
-    assertThat(StructuralKey.of(1234, VarIntCoder.of()).getKey(), equalTo(1234));
-    assertThat(StructuralKey.of("foobar", StringUtf8Coder.of()).getKey(), equalTo("foobar"));
+    Assert.assertThat(StructuralKey.of(1234, VarIntCoder.of()).getKey(), Matchers.equalTo(1234));
+    Assert.assertThat(
+        StructuralKey.of("foobar", StringUtf8Coder.of()).getKey(), Matchers.equalTo("foobar"));
     assertArrayEquals(StructuralKey.of(new byte[] {2, 9, -22}, ByteArrayCoder.of()).getKey(),
         new byte[] {2, 9, -22});
   }
@@ -51,21 +48,22 @@ public class StructuralKeyTest {
     byte[] original = new byte[] {1, 4, 9, 127, -22};
     StructuralKey<byte[]> key = StructuralKey.of(original, ByteArrayCoder.of());
 
-    assertThat(key.getKey(), not(theInstance(original)));
+    Assert.assertThat(key.getKey(), Matchers.not(Matchers.theInstance(original)));
   }
 
   @Test
   public void emptyKeysNotEqual() {
     StructuralKey<?> empty = StructuralKey.empty();
 
-    assertThat(empty, not(Matchers.<StructuralKey<?>>equalTo(StructuralKey.empty())));
-    assertThat(empty, Matchers.<StructuralKey<?>>equalTo(empty));
+    Assert.assertThat(
+        empty, Matchers.not(Matchers.<StructuralKey<?>>equalTo(StructuralKey.empty())));
+    Assert.assertThat(empty, Matchers.<StructuralKey<?>>equalTo(empty));
   }
 
   @Test
   public void objectEqualsTrueKeyEquals() {
     StructuralKey<Integer> original = StructuralKey.of(1234, VarIntCoder.of());
-    assertThat(StructuralKey.of(1234, VarIntCoder.of()), equalTo(original));
+    Assert.assertThat(StructuralKey.of(1234, VarIntCoder.of()), Matchers.equalTo(original));
   }
 
   @Test
@@ -75,7 +73,7 @@ public class StructuralKeyTest {
 
     StructuralKey<byte[]> otherKey =
         StructuralKey.of(new byte[] {1, 4, 9, 127, -22}, ByteArrayCoder.of());
-    assertThat(key, equalTo(otherKey));
+    Assert.assertThat(key, Matchers.equalTo(otherKey));
   }
 
   @Test
@@ -85,6 +83,6 @@ public class StructuralKeyTest {
 
     StructuralKey<byte[]> otherKey =
         StructuralKey.of(new byte[] {9, -128, 22}, ByteArrayCoder.of());
-    assertThat(key, not(equalTo(otherKey)));
+    Assert.assertThat(key, Matchers.not(Matchers.equalTo(otherKey)));
   }
 }
